requirements.txt
docopt
flask
flask-sqlalchemy
flask-migrate
flask_restful
Flask-Script
flask-admin
marshmallow-sqlalchemy
python-dotenv
sqlalchemy-json
gunicorn
requests
flask-jwt-extended
sqlalchemy-utils
babel
psycopg2-binary
geopy